+++ 06_LoopsOnline/src/loopsonline/PerfectNumberMethod.java
@@ -11,7 +11,7 @@ public class PerfectNumberMethod {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         System.out.println("Zadej cislo:");
-        int number = sc.nextInt();
+        long number = sc.nextLong();
         boolean perfect;
        
         perfect = isPerfect(number);
@@ -19,9 +19,9 @@ public class PerfectNumberMethod {
         System.out.println("Cislo " + ((perfect)?"je":"neni") + " dokonaly.");
     }
     
-    public static boolean isPerfect(int number){
-        int possibleDivisor = 2;
-        int sum = 1;
+    public static boolean isPerfect(long number){
+        long possibleDivisor = 2;
+        long sum = 1;
         while(possibleDivisor <= number/2){
             if(number % possibleDivisor == 0){
                 sum = sum + possibleDivisor;
